+++ otsclient/cmds.py
@@ -279,8 +279,11 @@ def upgrade_command(args):
             detached_timestamp = DetachedTimestampFile.deserialize(ctx)
 
         # IOError's are already handled by argparse
+        except BadMagicError:
+            logging.error("Error! %r is not a timestamp file." % old_stamp_fd.name)
+            sys.exit(1)
         except DeserializationError as exp:
-            logging.error("Invalid timestamp %r: %s" % (old_stamp_fd.name, exp))
+            logging.error("Invalid timestamp file %r: %s" % (old_stamp_fd.name, exp))
             sys.exit(1)
 
         changed = upgrade_timestamp(detached_timestamp.timestamp, args)
@@ -373,8 +376,11 @@ def verify_command(args):
     ctx = StreamDeserializationContext(args.timestamp_fd)
     try:
         detached_timestamp = DetachedTimestampFile.deserialize(ctx)
+    except BadMagicError:
+        logging.error("Error! %r is not a timestamp file." % args.timestamp_fd.name)
+        sys.exit(1)
     except DeserializationError as exp:
-        logging.error("Invalid timestamp %r: %s" % (args.timestamp_fd.name, exp))
+        logging.error("Invalid timestamp file %r: %s" % (args.timestamp_fd.name, exp))
         sys.exit(1)
 
     if args.hex_digest is not None:
@@ -417,8 +423,11 @@ def info_command(args):
     ctx = StreamDeserializationContext(args.file)
     try:
         detached_timestamp = DetachedTimestampFile.deserialize(ctx)
+    except BadMagicError:
+        logging.error("Error! %r is not a timestamp file." % args.file.name)
+        sys.exit(1)
     except DeserializationError as exp:
-        logging.error("Invalid timestamp %r: %s" % (args.file.name, exp))
+        logging.error("Invalid timestamp file %r: %s" % (args.file.name, exp))
         sys.exit(1)
 
     print("File %s hash: %s" % (detached_timestamp.file_hash_op.HASHLIB_NAME, hexlify(detached_timestamp.file_digest).decode('utf8')))